requirements.txt
gradio
requests
markdownify
smolagents==1.13.0
requests
duckduckgo_search
pandas
langchain-community
wikipedia-api
arxiv
numpy
pandas
matplotlib
Pillow
pytesseract
python-docx
openpyxl
tabula-py
pdf2image
faiss-cpu
sentence-transformers
pymupdf 
fitz
beautifulsoup4
pdfplumber
langchain-text-splitters
transformers
torch
